Allow to quit from the tray menu.
Fixes #6

Add a tray menu with a Quit action. This is not yet finished, because
currently, quitting makes qimaptray crash with a segmentation
fault. Actually, Qt tries to interrupt the imap thread and to destroy
the imap object. But the SSL objects cannot be freed because their
state has become corrupted when the thread was stopped.

This is to be fixed in a later commit.